(svn r17138) -Fix: some MSVC 64 bits compiler warnings

+++ src/settings_gui.cpp
@@ -615,7 +615,7 @@ public:
 
 				if (x >= 10) {
 					/* Increase button clicked */
-					val = min(val + sdb->interval, sdb->max);
+					val = min(val + sdb->interval, (int32)sdb->max);
 					this->clicked_increase = true;
 				} else {
 					/* Decrease button clicked */
